+++ openfisca_us/variables/irs/outputs.py
@@ -302,35 +302,6 @@ class c04470(Variable):
         return max_(0, tax_unit("c21060", period) - tax_unit("c21040", period))
 
 
-class exemption_phaseout_start(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "Exemption phaseout start"
-    definition_period = YEAR
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        return parameters(period).irs.income.exemption.phaseout.start[
-            tax_unit("mars", period)
-        ]
-
-
-class c04600(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Personal exemptions after phase-out"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        phaseout = parameters(period).irs.income.exemption.phaseout
-        phaseout_start = tax_unit("exemption_phaseout_start", period)
-        line_5 = max_(0, tax_unit("c00100", period) - phaseout_start)
-        line_6 = line_5 / (2500 / tax_unit("sep", period))
-        line_7 = phaseout.rate * line_6
-        return tax_unit("pre_c04600", period) * (1 - line_7)
-
-
 class c05700(Variable):
     value_type = float
     entity = TaxUnit
@@ -459,153 +430,6 @@ class c09200(Variable):
 income_tax_before_refundable_credits = variable_alias(
     "income_tax_before_refundable_credits", c09200
 )
-
-
-class c09600(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Alternative Minimum Tax"
-    unit = USD
-    documentation = "Alternative Minimum Tax (AMT) liability"
-
-    def formula(tax_unit, period, parameters):
-        c62100 = tax_unit("c62100", period)
-        # Form 6251, Part II top
-        amt = parameters(period).irs.income.amt
-        phaseout = amt.exemption.phaseout
-        mars = tax_unit("mars", period)
-        line29 = max_(
-            0,
-            (
-                amt.exemption.amount[mars]
-                - phaseout.rate * max_(0, c62100 - phaseout.start[mars])
-            ),
-        )
-        age_head = tax_unit("age_head", period)
-        child = amt.exemption.child
-        young_head = (age_head != 0) & (age_head < child.max_age)
-        no_or_young_spouse = tax_unit("age_spouse", period) < child.max_age
-        line29 = where(
-            young_head | no_or_young_spouse,
-            min_(line29, tax_unit("filer_earned", period) + child.amount),
-            line29,
-        )
-        line30 = max_(0, c62100 - line29)
-        brackets = amt.brackets
-        amount_over_threshold = line30 - brackets.thresholds["1"] / tax_unit(
-            "sep", period
-        )
-        line3163 = brackets.rates["1"] * line30 + brackets.rates["2"] * max_(
-            0, amount_over_threshold
-        )
-        dwks10, dwks13, dwks14, dwks19, e24515 = [
-            tax_unit(variable, period)
-            for variable in [
-                "dwks10",
-                "dwks13",
-                "dwks14",
-                "dwks19",
-                "filer_e24515",
-            ]
-        ]
-        form_6251_part_iii_required = np.any(
-            [
-                variable > 0
-                for variable in [
-                    dwks10,
-                    dwks13,
-                    dwks14,
-                    dwks19,
-                    e24515,
-                ]
-            ]
-        )
-
-        # Complete Form 6251, Part III
-
-        line37 = dwks13
-        line38 = e24515
-        line39 = min_(line37 + line38, dwks10)
-        line40 = min_(line30, line39)
-        line41 = max_(0, line30 - line40)
-        amount_over_threshold = max_(
-            0, line41 - amt.brackets.thresholds["1"] / tax_unit("sep", period)
-        )
-        line42 = (
-            amt.brackets.rates["1"] * line41
-            + amt.brackets.rates["2"] * amount_over_threshold
-        )
-        line44 = dwks14
-        cg = amt.capital_gains.brackets
-        line45 = max_(0, cg.thresholds["1"][mars] - line44)
-        line46 = min_(line30, line37)
-        line47 = min_(line45, line46)
-        cgtax1 = line47 * cg.rates["1"]
-        line48 = line46 - line47
-        line51 = dwks19
-        line52 = line45 + line51
-        line53 = max_(0, cg.thresholds["2"][mars] - line52)
-        line54 = min_(line48, line53)
-        cgtax2 = line54 * cg.rates["2"]
-        line56 = line47 + line54
-        line57 = where(line41 == line56, 0, line46 - line56)
-        linex2 = where(line41 == line56, 0, max_(0, line54 - line48))
-        cgtax3 = line57 * cg.rates["3"]
-        line61 = where(
-            line38 == 0,
-            0,
-            0.25 * max_(0, (line30 - line41 - line56 - line57 - linex2)),
-        )
-        line62 = line42 + cgtax1 + cgtax2 + cgtax3 + line61
-        line64 = min_(line3163, line62)
-        line31 = where(form_6251_part_iii_required, line64, line3163)
-        e07300 = tax_unit("e07300", period)
-
-        # Form 6251, Part II bottom
-        line32 = where(
-            tax_unit("f6251", period), tax_unit("filer_e62900", period), e07300
-        )
-        line33 = line31 - line32
-        return max_(
-            0,
-            line33
-            - max_(
-                0,
-                (
-                    tax_unit("taxbc", period)
-                    - e07300
-                    - tax_unit("c05700", period)
-                ),
-            ),
-        )
-
-
-alternative_minimum_tax = variable_alias("alternative_minimum_tax", c09600)
-
-
-class c10960(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = (
-        "American Opportunity Credit refundable amount from Form 8863"
-    )
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        return tax_unit("refundable_american_opportunity_credit", period)
-
-
-class c11070(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Child tax credit (refunded) from Form 8812"
-    unit = USD
-
-
-ctc_refund = variable_alias("ctc_refund", c11070)
 
 
 class c17000(Variable):
@@ -1043,22 +867,6 @@ class invinc_ec_base(Variable):
         return limited_capital_gain + other_inv_income
 
 
-class pre_c04600(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Personal exemption before phase-out"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        exemption = parameters(period).irs.income.exemption
-        return where(
-            tax_unit("dsi", period),
-            0,
-            tax_unit("xtot", period) * exemption.amount,
-        )
-
-
 class filer_setax(Variable):
     value_type = float
     entity = TaxUnit
@@ -1068,97 +876,6 @@ class filer_setax(Variable):
 
     def formula(tax_unit, period, parameters):
         return tax_unit_non_dep_sum("setax", tax_unit, period)
-
-
-class ymod(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "OASDI benefit tax variable"
-    documentation = "Variable that is used in OASDI benefit taxation logic"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        ymod2 = (
-            tax_unit("filer_e00400", period)
-            + (0.5 * tax_unit("filer_e02400", period))
-            - tax_unit("c02900", period)
-        )
-        YMOD3_ELEMENTS = ["filer_e03210", "filer_e03230", "filer_e03240"]
-        ymod3 = add(tax_unit, period, YMOD3_ELEMENTS)
-        return tax_unit("ymod1", period) + ymod2 + ymod3
-
-
-class ymod1(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "AGI increase"
-    documentation = "Variable that is included in AGI"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        DIRECT_INPUTS = [
-            "e00200",
-            "e00700",
-            "e00800",
-            "e01400",
-            "e01700",
-            "e02100",
-            "e02300",
-        ]
-        direct_inputs = add(
-            tax_unit,
-            period,
-            ["filer_" + variable for variable in DIRECT_INPUTS],
-        )
-        INVESTMENT_INCOME_SOURCES = [
-            "e00300",
-            "e00600",
-            "e01100",
-            "e01200",
-        ]
-        investment_income_sources = [
-            "filer_" + variable for variable in INVESTMENT_INCOME_SOURCES
-        ]
-        investment_income = add(
-            tax_unit, period, investment_income_sources
-        ) + tax_unit("c01000", period)
-        BUSINESS_INCOME_SOURCES = ["filer_e00900", "filer_e02000"]
-        business_income = add(tax_unit, period, BUSINESS_INCOME_SOURCES)
-        max_business_losses = parameters(
-            period
-        ).irs.ald.misc.max_business_losses[tax_unit("mars", period)]
-        business_income_losses_capped = max_(
-            business_income, -max_business_losses
-        )
-        return (
-            direct_inputs + investment_income + business_income_losses_capped
-        )
-
-
-class ubi(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Universal Basic Income benefit for filing unit"
-    unit = USD
-
-
-class taxable_ubi(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Amount of UBI benefit included in AGI"
-    unit = USD
-
-
-class nontaxable_ubi(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Amount of UBI benefit excluded from AGI"
-    unit = USD
 
 
 class aftertax_income(Variable):
